requirements (1).txt
streamlit
openai
requests
beautifulsoup4
Pillow
